+++ depthai_nodes/ml/messages/creators/__init__.py
@@ -1,11 +1,14 @@
-from .classification import create_classification_message
-from .classification_sequence import create_classification_sequence_message
+from .classification import (
+    create_classification_message,
+    create_classification_sequence_message,
+    create_multi_classification_message,
+)
 from .clusters import create_cluster_message
 from .detection import create_detection_message, create_line_detection_message
 from .image import create_image_message
 from .keypoints import create_hand_keypoints_message, create_keypoints_message
 from .map import create_map_message
-from .misc import create_age_gender_message, create_multi_classification_message
+from .misc import create_age_gender_message
 from .segmentation import create_sam_message, create_segmentation_message
 from .tracked_features import create_tracked_features_message
